# Conflicts:
#	LICENSE
#	src/mc/common/BuildInfo.h
#	src/mc/common/DimensionIdType.h
#	src/mc/common/Globals.h
#	src/mc/debug/PlayerRenderChunkModificationLatencyTracker.h
#	src/mc/deps/core/file/Path.h
#	src/mc/deps/core/file/PathPart.h
#	src/mc/deps/core/file/PathView.h
#	src/mc/deps/core/image/Image.h
#	src/mc/deps/core/sem_ver/SemVersionBase.h
#	src/mc/deps/core/string/HashedString.h
#	src/mc/deps/nbt/nbt/CompoundTag.h
#	src/mc/deps/nbt/nbt/CompoundTagVariant.h
#	src/mc/server/commands/CommandRegistry.h
#	src/mc/world/gamemode/InteractionResult.h
#	src/mc/world/inventory/network/ItemStackRequestSlotInfo.h
#	src/mc/world/level/BlockTickingQueue.h
#	src/mc/world/level/TickNextTickData.h
#	src/mc/world/level/block/actor/ChestBlockActor.h
#	src/mc/world/level/levelgen/structure/structurepools/StructurePoolBlockPredicateCappedRandomBlockReplacement.h

+++ src/mc/world/actor/ActorTypeNamespaceRules.h
@@ -2,7 +2,7 @@
 
 #include "mc/_HeaderOutputPredefine.h"
 
-enum class ActorTypeNamespaceRules : int {
+enum class ActorTypeNamespaceRules : uchar {
     ReturnWithoutNamespace = 0,
     ReturnWithNamespace    = 1,
 };
